api/requirements.txt
FastAPI
uvicorn
python-multipart
boto3
yt-dlp
motor
pydantic-settings